move get dataset tester
